added support for bar file export
