+++ modules/flowable-ui-modeler/flowable-ui-modeler-logic/src/main/java/org/flowable/app/service/editor/AppDefinitionExportService.java
@@ -16,37 +16,22 @@ import org.flowable.app.domain.editor.AppDefinition;
 import org.flowable.app.domain.editor.AppModelDefinition;
 import org.flowable.app.domain.editor.Model;
 import org.flowable.app.model.editor.AppDefinitionRepresentation;
-import org.flowable.app.repository.editor.ModelRepository;
-import org.flowable.app.security.SecurityUtils;
-import org.flowable.app.service.api.ModelService;
 import org.flowable.app.service.exception.BadRequestException;
 import org.flowable.app.service.exception.InternalServerErrorException;
 import org.flowable.editor.language.json.converter.BpmnJsonConverter;
 import org.flowable.editor.language.json.converter.util.CollectionUtils;
-import org.flowable.idm.api.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
 @Service
 @Transactional
-public class AppDefinitionExportService {
+public class AppDefinitionExportService extends BaseAppDefinitionService {
 
   private static final Logger logger = LoggerFactory.getLogger(AppDefinitionExportService.class);
-
-  @Autowired
-  protected ModelService modelService;
-
-  @Autowired
-  protected ModelRepository modelRepository;
-
-  @Autowired
-  protected ObjectMapper objectMapper;
   
   protected BpmnJsonConverter bpmnJsonConverter = new BpmnJsonConverter();
 
@@ -59,10 +44,22 @@ public class AppDefinitionExportService {
     Model appModel = modelService.getModel(modelId);
     AppDefinitionRepresentation appRepresentation = createAppDefinitionRepresentation(appModel);
 
-    createAppDefinitionZip(response, appModel, appRepresentation, SecurityUtils.getCurrentUserObject());
+    createAppDefinitionZip(response, appModel, appRepresentation);
   }
 
-  protected void createAppDefinitionZip(HttpServletResponse response, Model appModel, AppDefinitionRepresentation appDefinition, User user) {
+  public void exportDeployableAppDefinition(HttpServletResponse response, String modelId) throws IOException {
+
+    if (modelId == null) {
+      throw new BadRequestException("No application definition id provided");
+    }
+
+    Model appModel = modelService.getModel(modelId);
+    AppDefinitionRepresentation appRepresentation = createAppDefinitionRepresentation(appModel);
+
+    createAppDefinitionBar(response, appModel, appRepresentation);
+  }
+
+  protected void createAppDefinitionZip(HttpServletResponse response, Model appModel, AppDefinitionRepresentation appDefinition) {
     response.setHeader("Content-Disposition", "attachment; filename=" + appDefinition.getName() + ".zip");
     try {
       ServletOutputStream servletOutputStream = response.getOutputStream();
@@ -74,8 +71,8 @@ public class AppDefinitionExportService {
 
       List<AppModelDefinition> modelDefinitions = appDefinition.getDefinition().getModels();
       if (CollectionUtils.isNotEmpty(modelDefinitions)) {
-        Map<String, Model> formMap = new HashMap<String, Model>();
-        Map<String, Model> decisionTableMap = new HashMap<String, Model>();
+        Map<String, Model> formMap = new HashMap<>();
+        Map<String, Model> decisionTableMap = new HashMap<>();
 
         for (AppModelDefinition modelDef : modelDefinitions) {
           Model model = modelService.getModel(modelDef.getId());
@@ -113,6 +110,26 @@ public class AppDefinitionExportService {
       throw new InternalServerErrorException("Could not generate app definition zip archive");
     }
   }
+
+  public void createAppDefinitionBar(HttpServletResponse response, Model appModel, AppDefinitionRepresentation appDefinition) {
+      response.setHeader("Content-Disposition", "attachment; filename=" + appDefinition.getName() + ".bar");
+
+      try {
+        byte[] deployZipArtifact = createDeployableZipArtifact(appModel, appDefinition.getDefinition());
+
+        ServletOutputStream servletOutputStream = response.getOutputStream();
+        response.setContentType("application/zip");
+        servletOutputStream.write(deployZipArtifact);
+
+        // Flush and close stream
+        servletOutputStream.flush();
+        servletOutputStream.close();
+
+      } catch (Exception e) {
+        logger.error("Could not generate app definition bar archive", e);
+        throw new InternalServerErrorException("Could not generate app definition bar archive");
+      }
+    }
   
   protected void createZipEntries(Model model, String directoryName, ZipOutputStream zipOutputStream) throws Exception {
     createZipEntry(zipOutputStream, directoryName + "/" + model.getKey() + ".json", createModelEntryJson(model));
@@ -163,4 +180,5 @@ public class AppDefinitionExportService {
     zipOutputStream.write(content);
     zipOutputStream.closeEntry();
   }
+
 }
